Tidying and Additions to PlaiCraft Project Analysis
Added one more visualization in the form of a line graph. Fixed the scale of the scatter plot. Started the paragraph for the methods and plans section of the project planning. Fixed any spelling and grammar errors and tried to make it more concise. Also, added headers for each visualization.